fix c++ warning

+++ src/apps/msg_gen_tcp/main.cc
@@ -227,7 +227,7 @@ void ClientSendOne(ThreadCtx *thread_ctx, uint64_t window_slot) {
 
   const int ret = write(thread_ctx->sock_fd, thread_ctx->tx_message.data(),
                         thread_ctx->tx_message.size());
-  if (ret == FLAGS_tx_msg_size) {
+  if (ret == static_cast<int>(FLAGS_tx_msg_size)) {
     stats_cur.tx_success++;
     stats_cur.tx_bytes += FLAGS_tx_msg_size;
   } else {
